Rename bucket UI components
Rename backlog bucket UI classes to shorter bucket-based names
while keeping the underlying model, routes, params, and user-
facing copy unchanged.

This keeps the UI layer less wordy without widening the rename
into persistence and service code.
