+++ modules/backlogs/app/components/backlogs/bucket_component.rb
@@ -29,34 +29,49 @@
 #++
 
 module Backlogs
-  class NewBacklogBucketFormComponent < ApplicationComponent
-    include ApplicationHelper
+  class BucketComponent < ApplicationComponent
+    include Primer::AttributesHelper
     include OpTurbo::Streamable
-    include OpPrimer::ComponentHelpers
+    include Backlogs::CommonHelper
 
-    FORM_ID = NewBacklogBucketDialogComponent::FORM_ID
+    attr_reader :backlog_bucket, :project, :work_packages, :current_user
 
-    attr_reader :backlog_bucket
-
-    def initialize(backlog_bucket:, base_errors: nil)
-      super
+    def initialize(backlog_bucket:, project:, current_user: User.current, **system_arguments)
+      super()
 
       @backlog_bucket = backlog_bucket
-      @base_errors = base_errors
+      @project = project
+      @current_user = current_user
+      @work_packages = backlog_bucket.work_packages
+
+      @system_arguments = system_arguments
+      @system_arguments[:id] = dom_id(backlog_bucket)
+      @system_arguments[:list_id] = "#{@system_arguments[:id]}-list"
+      @system_arguments[:padding] = :condensed
+      @system_arguments[:data] = merge_data(
+        @system_arguments,
+        { data: drop_target_config },
+        { data: { test_selector: "backlog-bucket-#{backlog_bucket.id}" } }
+      )
+    end
+
+    def wrapper_uniq_by
+      backlog_bucket.id
     end
 
     private
 
-    def http_verb
-      @backlog_bucket.new_record? ? :post : :put
+    def folded?
+      current_user.pref[:backlogs_versions_default_fold_state] == "closed"
     end
 
-    def form_url
-      if @backlog_bucket.new_record?
-        project_backlogs_backlog_buckets_path(@backlog_bucket.project)
-      else
-        project_backlogs_backlog_bucket_path(@backlog_bucket.project, @backlog_bucket)
-      end
+    def drop_target_config
+      {
+        generic_drag_and_drop_target: "container mirrorContainer",
+        target_container_accessor: ":scope > ul",
+        target_id: backlog_bucket.persisted? ? "backlog_bucket:#{backlog_bucket.id}" : "inbox",
+        target_allowed_drag_type: "story"
+      }
     end
   end
 end
