+++ modules/backlogs/app/components/backlogs/bucket_header_component.html.erb
@@ -70,7 +70,7 @@ See COPYRIGHT and LICENSE files for more details.
     <% end %>
 
     <% grid.with_area(:menu) do %>
-      <%= render(Backlogs::BacklogBucketMenuComponent.new(backlog_bucket:, project:)) %>
+      <%= render(Backlogs::BucketMenuComponent.new(backlog_bucket:, project:)) %>
     <% end %>
   <% end %>
 <% end %>
