+++ modules/backlogs/app/components/backlogs/new_bucket_dialog_component.html.erb
@@ -28,23 +28,40 @@ See COPYRIGHT and LICENSE files for more details.
 ++#%>
 
 <%=
-  component_wrapper do
-    primer_form_with(
-      id: Backlogs::NewBacklogBucketFormComponent::FORM_ID,
-      model: backlog_bucket,
-      scope: :backlog_bucket,
-      method: http_verb,
-      class: "op-backlog-buckets--form",
-      url: form_url
-    ) do |f|
-      flex_layout(mb: 2) do |flex|
-        if @base_errors&.any?
-          flex.with_row do
-            render(Primer::Alpha::Banner.new(mb: 3, icon: :stop, scheme: :danger)) { @base_errors.join("\n") }
-          end
+  render(
+    Primer::Alpha::Dialog.new(
+      title:,
+      size: :large,
+      id: DIALOG_ID
+    )
+  ) do |d|
+    d.with_header(variant: :large)
+
+    d.with_body do
+      render(Backlogs::NewBucketFormComponent.new(backlog_bucket:))
+    end
+
+    d.with_footer do
+      component_collection do |buttons|
+        buttons.with_component(
+          Primer::Beta::Button.new(
+            data: {
+              close_dialog_id: DIALOG_ID
+            }
+          )
+        ) do
+          t("button_cancel")
         end
-        flex.with_row(mb: 3) do
-          render Backlogs::BacklogBuckets::DetailsForm.new(f)
+
+        buttons.with_component(
+          Primer::Beta::Button.new(
+            scheme: :primary,
+            form: FORM_ID,
+            data: { turbo: true },
+            type: :submit
+          )
+        ) do
+          button_caption
         end
       end
     end
