+++ modules/backlogs/app/components/backlogs/new_bucket_dialog_component.rb
@@ -29,49 +29,37 @@
 #++
 
 module Backlogs
-  class BacklogBucketComponent < ApplicationComponent
-    include Primer::AttributesHelper
+  class NewBucketDialogComponent < ApplicationComponent
     include OpTurbo::Streamable
-    include Backlogs::CommonHelper
+    include OpPrimer::ComponentHelpers
+    include Primer::FetchOrFallbackHelper
 
-    attr_reader :backlog_bucket, :project, :work_packages, :current_user
+    DIALOG_ID = "new-backlog-bucket-dialog"
+    FORM_ID = "new-backlog-bucket-dialog-form"
+    FOOTER_ID = "new-backlog-bucket-dialog-footer"
 
-    def initialize(backlog_bucket:, project:, current_user: User.current, **system_arguments)
-      super()
+    STATE_DEFAULT = :create
+    STATE_OPTIONS = [STATE_DEFAULT, :edit].freeze
+
+    attr_reader :backlog_bucket, :state
+
+    delegate :create?, :edit?, to: :state
+
+    def initialize(backlog_bucket:, state: STATE_DEFAULT)
+      super
 
       @backlog_bucket = backlog_bucket
-      @project = project
-      @current_user = current_user
-      @work_packages = backlog_bucket.work_packages
-
-      @system_arguments = system_arguments
-      @system_arguments[:id] = dom_id(backlog_bucket)
-      @system_arguments[:list_id] = "#{@system_arguments[:id]}-list"
-      @system_arguments[:padding] = :condensed
-      @system_arguments[:data] = merge_data(
-        @system_arguments,
-        { data: drop_target_config },
-        { data: { test_selector: "backlog-bucket-#{backlog_bucket.id}" } }
-      )
-    end
-
-    def wrapper_uniq_by
-      backlog_bucket.id
+      @state = ActiveSupport::StringInquirer.new(fetch_or_fallback(STATE_OPTIONS, state, STATE_DEFAULT).to_s)
     end
 
     private
 
-    def folded?
-      current_user.pref[:backlogs_versions_default_fold_state] == "closed"
+    def title
+      create? ? t(:label_backlog_bucket_new) : t(:label_backlog_bucket_edit)
     end
 
-    def drop_target_config
-      {
-        generic_drag_and_drop_target: "container mirrorContainer",
-        target_container_accessor: ":scope > ul",
-        target_id: backlog_bucket.persisted? ? "backlog_bucket:#{backlog_bucket.id}" : "inbox",
-        target_allowed_drag_type: "story"
-      }
+    def button_caption
+      create? ? t(:button_create) : t(:button_save)
     end
   end
 end
