+++ modules/backlogs/config/locales/en.yml
@@ -130,11 +130,11 @@ en:
       title: "Backlog admin settings are evolving"
       text: "We are currently redesigning the Backlogs module. Admin settings for sprints and backlogs will be visible here in the near future. Project-level settings remain available."
 
-    backlog_bucket_component:
+    bucket_component:
       blankslate_title: "Backlog bucket is empty"
       blankslate_description: "Drag items here to add them."
 
-    backlog_bucket_item_component:
+    bucket_item_component:
       label_actions: "Work package actions"
 
     sprint_component:
@@ -173,13 +173,13 @@ en:
         receive_shared_no_actions_description_text: "This project receives shared sprints from a different project, but none are available right now."
         settings_link_text: "project settings"
 
-    backlog_bucket_menu_component:
+    bucket_menu_component:
       label_actions: "Backlog bucket actions"
       action_menu:
         edit_backlog_bucket: "Edit backlog bucket"
         delete_backlog_bucket: "Delete backlog bucket"
 
-    backlog_bucket_header_component:
+    bucket_header_component:
       label_work_package_count:
         zero: "No stories in backlog bucket"
         one: "%{count} story in backlog bucket"
